+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Added
 - drel.core.builders.BaseFullRequestLogBuilder: user can be passed into dunder call method
+- drel.core.schemas.FullRequestLogSchema: stats - dict with request stats (like duration)
+- drel.requests.api.log: duration kwarg
+- drel.django.api.LoggingMiddleware: duration logging 
 
 ### Fixed
 - drel.django.LoggingMiddleware: log user after request processing